batch: Move ownership unit types
The ownership unit module builds display-derived units, validates selected
units, filters them, and rebuilds ownership metadata.

The value objects for those operations still live beside the scanning and
mutation helpers. That makes the operation module define the record shape
that callers may need independently.

This commit addresses that split by moving OwnershipUnit and
OwnershipUnitKind into ownership_unit_types. Ownership unit operations now
use role-named local aliases for those records.

A following tests commit will move focused imports and guard the
value-object boundary.
